ссылки на self в замыканиях- weak. Массив questions перенесен из MovieQuizViewController в class QuestionFactory. Структуры QuizStepViewModel, QuizResultsViewModel и QuizQuestion перенесены из MovieQuizViewController в соответствующие файлы
